refactor extension function to match with ktlint rule
